For the most part, now SeeFloorWithBadBlocks is isolated and is not used anywhere in production code - there is just one test class
